Changed some arrays back to single precision

+++ erdst/realcal.F90
@@ -21,8 +21,8 @@ module realcal
   implicit none
   
   ! "straight" coordinate system
-  real(kind=8), allocatable :: sitepos_normal(:, :)
-  real(kind=8) :: cell_normal(3, 3), invcell_normal(3), cell_len_normal(3)
+  real, allocatable :: sitepos_normal(:, :)
+  real :: cell_normal(3, 3), invcell_normal(3), cell_len_normal(3)
   logical :: is_cuboid
   real, parameter :: check_rotate = 1e-8, cuboid_thres = 1e-8
   real, parameter :: cutoff_thres = 1e-4
